Added TotalHoursChart component to show client total hours
